[FEATURE] Implemented main contacts functions

+++ app/src/main/java/com/sudox/android/common/repository/main/ContactsRepository.kt
@@ -14,15 +14,13 @@ class ContactsRepository(private val protocolClient: ProtocolClient,
                          private val contactsDao: ContactsDao) {
 
     val contactsLoadLiveData = MutableLiveData<List<Contact>>()
-    val contactsUpdateLiveData = MutableLiveData<State>()
 
     fun initContactsListeners() {
-
-        protocolClient.listenMessage("contacts.add", object : ResponseCallback<ContactAddDTO> {
-            override fun onMessage(response: ContactAddDTO) {
+        protocolClient.listenMessage("contacts.add", object : ResponseCallback<ContactAddRemoveDTO> {
+            override fun onMessage(response: ContactAddRemoveDTO) {
                 if (response.code != 0) {
                     val usersGetDTO = UsersGetDTO()
-                    usersGetDTO.id = response.aid
+                    usersGetDTO.id = response.id
 
                     protocolClient.makeRequest("users.get", usersGetDTO, object : ResponseCallback<UsersGetDTO> {
                         override fun onMessage(response: UsersGetDTO) {
@@ -43,18 +41,19 @@ class ContactsRepository(private val protocolClient: ProtocolClient,
             }
         })
 
-        protocolClient.listenMessage("contacts.remove", object : ResponseCallback<ContactDTO> {
-            override fun onMessage(response: ContactDTO) {
-                contactsDao.deleteContactById(response.id)
+        protocolClient.listenMessage("contacts.remove", object : ResponseCallback<ContactAddRemoveDTO> {
+            override fun onMessage(response: ContactAddRemoveDTO) {
+                if(response.code != 0) {
+                    contactsDao.deleteContactById(response.id)
+                    requestAllContactsFromDB()
+                }
             }
         })
     }
 
     fun getAllContactsFromServer(): Single<State> = Single.unsafeCreate {
-        val contactsGetDTO = ContactsGetDTO()
-        contactsGetDTO.count = 10
 
-        protocolClient.makeRequest("contacts.get", contactsGetDTO, object : ResponseCallback<ContactsGetDTO> {
+        protocolClient.makeRequest("contacts.get", SimpleAnswerDTO(), object : ResponseCallback<ContactsGetDTO> {
             override fun onMessage(response: ContactsGetDTO) {
                 if (response.code != 0) {
                     for (i in 0..(response.items!!.length() - 1)) {
@@ -102,15 +101,23 @@ class ContactsRepository(private val protocolClient: ProtocolClient,
     }
 
     fun contactAdd(id: String) {
-        val contactAddDTO = ContactAddDTO()
-        contactAddDTO.id = id
+        val contactAddDTO = ContactAddRemoveDTO()
+        contactAddDTO.sendId = id
         protocolClient.sendMessage("contacts.add", contactAddDTO)
     }
 
+    fun removeContact(id: String) {
+        val contactRemove = ContactAddRemoveDTO()
+        contactRemove.sendId = id
+        protocolClient.sendMessage("contacts.remove", contactRemove)
+    }
+
     fun requestAllContactsFromDB() {
-//        //TODO: try to find the best way
+        //TODO: try to find the best way
         Schedulers.io().scheduleDirect {
             contactsLoadLiveData.postValue(contactsDao.getContacts())
         }
     }
+
+
 }
